Splits the lib.rs file
